Update Iodine::JSON logic and documentation
