+++ ext/iodine/fiobj_hash.c
@@ -126,6 +126,7 @@ size_t fiobj_hash_count(const FIOBJ o) {
   assert(o && FIOBJ_TYPE_IS(o, FIOBJ_T_HASH));
   return fio_hash_count(&obj2hash(o)->hash);
 }
+intptr_t fiobj_hash2num(const FIOBJ o) { return (intptr_t)fiobj_hash_count(o); }
 static size_t fiobj_hash_is_true(const FIOBJ o) {
   return fiobj_hash_count(o) != 0;
 }
@@ -142,7 +143,7 @@ const fiobj_object_vtable_s FIOBJECT_VTABLE_HASH = {
     .each = fiobj_hash_each1,
     .is_true = fiobj_hash_is_true,
     .to_str = fiobject___noop_to_str,
-    .to_i = fiobject___noop_to_i,
+    .to_i = fiobj_hash2num,
     .to_f = fiobject___noop_to_f,
 };
